requirements.txt
autopep8==1.6.0
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-Cors==3.0.10
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysql-connector-python==8.1.0
numpy==1.26.0
pandas==2.1.1
pip==23.2.1
protobuf==4.21.12
psycopg2==2.9.3
pycodestyle==2.8.0
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2023.3.post1
setuptools==68.2.2
six==1.16.0
toml==0.10.2
tzdata==2023.3
Werkzeug==2.1.1
wheel==0.41.2
zipp==3.8.0